+++ database/migrations/1000_create_product_types_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->timestamps();
             $table->string('name');
             $table->string('code');
-            $table->float('volume');
+            $table->decimal('volume');
             $table->string('dimension');
 
             $table->unsignedBigInteger('product');
